Move Project files into their own folder
